modify schema naming for postgre

+++ src/main/java/win/doyto/ormcamparison/dq/salary/SalaryController.java
@@ -6,5 +6,5 @@ import win.doyto.query.web.controller.AbstractEIQController;
 
 @RestController
 @RequestMapping("/dq/salary")
-public class SalaryController extends AbstractEIQController<SalaryEntity, Long, SalaryQuery> {
+public class SalaryController extends AbstractEIQController<SalaryEntity, Integer, SalaryQuery> {
 }
