+++ src/main/java/win/doyto/ormcamparison/dq/salary/SalaryEntity.java
@@ -3,15 +3,17 @@ package win.doyto.ormcamparison.dq.salary;
 import lombok.Data;
 import win.doyto.query.entity.AbstractPersistable;
 
+import java.math.BigDecimal;
+
 @Data
-public class SalaryEntity extends AbstractPersistable<Long> {
+public class SalaryEntity extends AbstractPersistable<Integer> {
     private Integer workYear;
     private String experienceLevel;
     private String employmentType;
     private String jobTitle;
-    private Double salary;
+    private BigDecimal salary;
     private String salaryCurrency;
-    private Double salaryInUsd;
+    private BigDecimal salaryInUsd;
     private String employeeResidence;
     private Integer remoteRatio;
     private String companyLocation;
