+++ src/main/java/win/doyto/ormcamparison/jdbc/salary/SalaryEntity.java
@@ -2,16 +2,18 @@ package win.doyto.ormcamparison.jdbc.salary;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class SalaryEntity {
-    private Long id;
+    private Integer id;
     private Integer workYear;
     private String experienceLevel;
     private String employmentType;
     private String jobTitle;
-    private Double salary;
+    private BigDecimal salary;
     private String salaryCurrency;
-    private Double salaryInUsd;
+    private BigDecimal salaryInUsd;
     private String employeeResidence;
     private Integer remoteRatio;
     private String companyLocation;
